Add need_compilation argument
If this is set to FALSE, then we do not check for the compiler,
and we also do not compile Rcpp attributes. pkgdepends can set
this, based on the `NeedsCompilation` entry in the package
database.

By default it is NA, meaning that it is detected via the presence
of an src directory.