Support native routine registration for C
pkgbuild now supports generation of the native routine registration for
C code. The way this works is it uses
tools::package_native_routines_registration_skeleton() to generate the
registrations, than inserts them in `src/init.c` between pkgdown
specific comments.

The rest of the file is passed through unchanged, so if you need to have
custom init code or additional includes you can do so just by adding it
to the same file, only the code
within the comments will be altered when the registration is updated.

Fixes #50