Rtools tests (#201)
* Rtools tests: use RSPM

So we install everything from binaries.

* Rtools test: remove pre-installed R

So it does not bother us.

* Install Rtools first

So we can install the dependencies of pkgbuild. Then we'll delete it
later.

* Remove preinstalled Rtools as well

* Debug shell

* Install pkgbuild with base R

pak is currently broken for R-alpha and R-devel.

* No need to install pandoc

Because of an r-lib/actions bug the pandoc check
fails if some soft dependencies cannot be installed
on this R version, so it fails on R 3.6.x.

* Add some actual pkgbuild tests

* Try with jpeg instead of tiff

Maybe tiff is just broken on R 4.2.x?

* Fox repos config

* Test w/o pkgbuild as well

* Speed up removing R & rtools

Rename them and clean the registry, that should be enough.

* Work around a bug

Hopefully unrelated to us. Somebody should set PKG_CONFIG_PATH, though.