Fix binary build of package files with non-standard names
No need to run `bootstrap.R` when building binary
packages, i.e. when running `R CMD INSTALL`.

This uses `get_desc_config_flag()`, which does not work
on package files with non-standard names, probably a
bug in the desc package. (It is also pretty slow because it
involves an extra uncompression.)

Closes #208.